% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_VarghaDelaney.R
\name{A_VarghaDelaney}
\alias{A_VarghaDelaney}
\title{Vargha & Delaney's A}
\usage{
A_VarghaDelaney(
  control,
  experimental,
  bootstrap = NULL,
  conf.level = 0.95,
  warn = FALSE
)
}
\arguments{
\item{control}{A vector with the data for the control condition.}

\item{experimental}{A vector with the data from the experimental condition.}

\item{bootstrap}{The number of bootstrap samples to use to compute
confidence intervals, or NULL to not compute confidence intervals.}

\item{conf.level}{The confidence level of the confidence intervals.}

\item{warn}{Whether to allow the \code{\link[stats:wilcox.test]{stats::wilcox.test()}} function to emit
warnings, for example if ties are encountered.}
}
\value{
A numeric vector of length 1 with the A value, named 'A'.
}
\description{
Vargha & Delaney's A
}
\examples{
ufs::A_VarghaDelaney(1:8, 3:12);
}
