% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipedjmv.h.R
\name{aipedjmv}
\alias{aipedjmv}
\title{Sample size for accuracy: d}
\usage{
aipedjmv(d = 0.5, w = 0.1, conf.level = 95)
}
\arguments{
\item{d}{.}

\item{w}{.}

\item{conf.level}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a html \cr
\code{results$aipePlot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Sample size for accuracy: d
}
