% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiperjmv.h.R
\name{aiperjmv}
\alias{aiperjmv}
\title{Sample size for accuracy: r}
\usage{
aiperjmv(r = 0.3, w = 0.1, conf.level = 95)
}
\arguments{
\item{r}{.}

\item{w}{.}

\item{conf.level}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a html \cr
\code{results$aipePlot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Sample size for accuracy: r
}
