% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areColors.R
\name{areColors}
\alias{areColors}
\title{Check whether elements of a vector are valid colors}
\usage{
areColors(x)
}
\arguments{
\item{x}{The vector.}
}
\value{
A logical vector.
}
\description{
This function by Josh O'Brien checks whether elements of a vector are valid
colors. It has been copied from a Stack Exchange answer (see
\url{https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation}).
}
\examples{

ufs::areColors(c(NA, "black", "blackk", "1", "#00", "#000000"));

}
\author{
Josh O'Brien

Maintainer: Gjalt-Jorn Peters \href{mailto:ufs@opens.science}{ufs@opens.science}
}
\keyword{utilities}
