% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attenuate.r.R
\name{attenuate.r}
\alias{attenuate.r}
\title{Attenuate a Pearson's r estimate for unreliability in the measurements}
\usage{
attenuate.r(r, reliability1, reliability2)
}
\arguments{
\item{r}{The (disattenuated) value of Pearson's r}

\item{reliability1, reliability2}{The reliabilities of the two variables}
}
\value{
The attenuated value of Pearson's r
}
\description{
Attenuate a Pearson's r estimate for unreliability in the measurements
}
\examples{
attenuate.r(.5, .8, .9);
}
