% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfi-data.R
\docType{data}
\name{bfi-data}
\alias{bfi-data}
\alias{bfi}
\title{25 Personality items representing 5 factors}
\format{
A \code{data.frame} with 2800 rows and 28 columns.
}
\usage{
data(bfi)
}
\description{
This is a dataset lifted from the \code{psychTools} package (which was
originally in the \code{psych} package). For details, please check
that help page (using "\code{psychTools::bfi}").
}
\examples{
data(bfi);
}
\keyword{data}
