% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confintdjmv.h.R
\name{confintdjmv}
\alias{confintdjmv}
\title{Effect Size Confidence Interval: Cohens's d}
\usage{
confintdjmv(d = 0.5, n = 128, conf.level = 95)
}
\arguments{
\item{d}{.}

\item{n}{.}

\item{conf.level}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a html \cr
\code{results$ciPlot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Effect Size Confidence Interval: Cohens's d
}
