% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confintrjmv.h.R
\name{confintrjmv}
\alias{confintrjmv}
\title{Effect Size Confidence Interval: Pearson's r}
\usage{
confintrjmv(r = 0.3, N = 400, conf.level = 95)
}
\arguments{
\item{r}{.}

\item{N}{.}

\item{conf.level}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a html \cr
\code{results$ciPlot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Effect Size Confidence Interval: Pearson's r
}
