% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.d.to.nnc.R
\name{convert.d.to.nnc}
\alias{convert.d.to.nnc}
\alias{convert.d.to.eer}
\title{Helper functions for Numbers Needed for Change}
\usage{
convert.d.to.nnc(d, cer, r = 1, eventDesirable = TRUE,
  eventIfHigher = TRUE)
}
\arguments{
\item{d}{The value of Cohen's \emph{d}.}

\item{cer}{The Control Event Rate.}

\item{r}{The correlation between the determinant and behavior (for mediated
Numbers Needed for Change).}

\item{eventDesirable}{Whether an event is desirable or undesirable.}

\item{eventIfHigher}{Whether scores above or below the threshold are
considered 'an event'.}
}
\value{
The converted value.
}
\description{
These two functions are used by \code{\link{nnc}} to compute the Numbers
Needed for Change.
}
\details{
These two functions are used by \code{\link{nnc}} to compute the Numbers
Needed for Change.
}
\examples{

convert.d.to.eer(d=.5, cer=.25);
convert.d.to.nnc(d=.5, cer=.25);

}
\references{
Gruijters, S. L. K., & Peters, G.-J. Y. (2017). Introducing the
Numbers Needed for Change (NNC): A practical measure of effect size for
intervention research.
}
\seealso{
\code{\link{nnc}}
}
\author{
Gjalt-Jorn Peters & Stefan Gruijters

Maintainer: Gjalt-Jorn Peters <gjalt-jorn@userfriendlyscience.com>
}
\keyword{utilities}
