% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.d.to.U3.R
\name{convert.d.to.U3}
\alias{convert.d.to.U3}
\title{Convert Cohen's d to U3}
\usage{
convert.d.to.U3(d)
}
\arguments{
\item{d}{Cohen's d.}
}
\value{
An unnames numeric vector with the U3 values.
}
\description{
This function simply returns the result of \code{\link[=pnorm]{pnorm()}} for
Cohen's d.
}
\examples{
convert.d.to.U3(.5);
}
