% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFS.R
\name{convert_color}
\alias{convert_color}
\title{Convert Color}
\usage{
convert_color(mag = NULL, gradient = "RdYlGn", normalize = TRUE)
}
\arguments{
\item{mag}{The magnitudes to convert}

\item{gradient}{The gradient to use for coloring. "~" reverses gradient direction. Defaults to "RdYlGn".}

\item{normalize}{Normalizes the magnitudes to 1-9.9 before color casting. Defaults to TRUE.}
}
\value{
Returns a vector containing the color magnitudes. Has a side-effect that changes the color palette.
}
\description{
Casts a set of magnitudes to a color gradient.
}
\examples{
convert_color(party)
}
\keyword{casting}
\keyword{color,}
