% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disattenuate.d.R
\name{disattenuate.d}
\alias{disattenuate.d}
\title{Disattentuate a Cohen's d estimate for unreliability in the continuous variable}
\usage{
disattenuate.d(d, reliability)
}
\arguments{
\item{d}{The (attenuated) value of Cohen's d}

\item{reliability}{The reliability of the measurements of the continuous variable}
}
\value{
The disattenuated value of Cohen's d
}
\description{
Disattentuate a Cohen's d estimate for unreliability in the continuous variable
}
\examples{
disattenuate.d(.5, .8);
}
