% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disattenuate.r.R
\name{disattenuate.r}
\alias{disattenuate.r}
\title{Disattentuate a Pearson's r estimate for unreliability}
\usage{
disattenuate.r(r, reliability1, reliability2)
}
\arguments{
\item{r}{The (attenuated) value of Pearson's r}

\item{reliability1, reliability2}{The reliabilities of the two variables}
}
\value{
The disattenuated value of Pearson's r
}
\description{
Disattentuate a Pearson's r estimate for unreliability
}
\examples{
disattenuate.r(.5, .8, .9);
}
