% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractVarName.R
\name{extractVarName}
\alias{extractVarName}
\title{Extract variable names}
\usage{
extractVarName(x)
}
\arguments{
\item{x}{A character vector of one or more variable names.}
}
\value{
The actual variables name, with all containing objectes stripped
off.
}
\description{
Functions often get passed variables from within dataframes or other lists.
However, printing these names with all their dollar signs isn't very
userfriendly. This function simply uses a regular expression to extract the
actual name.
}
\examples{

extractVarName('mtcars$mpg');

}
\author{
Gjalt-Jorn Peters

Maintainer: Gjalt-Jorn Peters \href{mailto:ufs@opens.science}{ufs@opens.science}
}
\keyword{utils}
