% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_failsafe.R
\name{fa_failsafe}
\alias{fa_failsafe}
\title{Do factor-analysis, logging warnings and errors}
\usage{
fa_failsafe(
  ...,
  n.repeatOnWarning = 50,
  warningTolerance = 2,
  silentRepeatOnWarning = FALSE,
  showWarnings = TRUE
)
}
\arguments{
\item{...}{The arguments for \code{fa} in \code{psych}.}

\item{n.repeatOnWarning}{How often to repeat on warnings (in the
hopes of getting a run without warnings).}

\item{warningTolerance}{How many warnings are accepted.}

\item{silentRepeatOnWarning}{Whether to be chatty or silent when
repeating after warnings.}

\item{showWarnings}{Whether to show the warnings.}
}
\value{
A list with the \code{fa} object and a \code{warnings} and an \code{errors} object.
}
\description{
Do factor-analysis, logging warnings and errors
}
