% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBarChart.R
\name{ggBarChart}
\alias{ggBarChart}
\title{Bar chart using ggplot}
\usage{
ggBarChart(vector, plotTheme = ggplot2::theme_bw(), ...)
}
\arguments{
\item{vector}{The vector to display in the bar chart.}

\item{plotTheme}{The theme to apply.}

\item{\dots}{And additional arguments are passed to \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} plot is returned.
}
\description{
This function provides a simple interface to create a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
bar chart.
}
\examples{

ggBarChart(mtcars$cyl);

}
\seealso{
\code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}
}
\author{
Gjalt-Jorn Peters

Maintainer: Gjalt-Jorn Peters \href{mailto:ufs@opens.science}{ufs@opens.science}
}
