% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFS.R
\name{gvm}
\alias{gvm}
\title{Greedy Variance Maximization}
\usage{
gvm(df, n, show = TRUE, ...)
}
\arguments{
\item{df}{The dataframe to plot.}

\item{show}{Displays the plot. Defaults to TRUE.}

\item{num_vectors}{The number of vectors to keep}
}
\value{
\item{indices}{The indices of the most important vectors.}

\item{var_explained}{The variance explained by each of the vectors.}
}
\description{
Runs GVM and plots the variance explained by a set of column vectors.
}
\examples{
gvm(senate, 2, main="Bill GVM Variance Explained")
gvm(t(senate), 2, main="Senator GVM Variance Explained")
}
\keyword{gvm}
