% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelseObj.R
\name{ifelseObj}
\alias{ifelseObj}
\title{Conditional returning of an object}
\usage{
ifelseObj(condition, ifTrue, ifFalse)
}
\arguments{
\item{condition}{Condition to evaluate.}

\item{ifTrue}{Object to return if the condition is true.}

\item{ifFalse}{Object to return if the condition is false.}
}
\value{
One of the two objects
}
\description{
The ifelseObj function just evaluates a condition, returning
  one object if it's true, and another if it's false.
}
\examples{
dat <- ifelseObj(sample(c(TRUE, FALSE), 1), mtcars, Orange);

}
