% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nr.R
\name{is.nr}
\alias{is.nr}
\title{\code{NULL} and \code{NA} 'proof' checking of whether something is a number}
\usage{
is.nr(x)
}
\arguments{
\item{x}{The value or vector to check.}
}
\value{
TRUE or FALSE.
}
\description{
Convenience function that returns TRUE if the argument is not null, not NA,
and is.numeric.
}
\examples{
is.nr(8);    ### Returns TRUE
is.nr(NULL); ### Returns FALSE
is.nr(NA);   ### Returns FALSE

}
