% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTrue.R
\name{isTrue}
\alias{isTrue}
\title{More flexible version of isTRUE}
\usage{
isTrue(x, na = FALSE)
}
\arguments{
\item{x}{The vector to check for \code{TRUE}, \code{FALSE}, and \code{NA} values.}

\item{na}{What to return for \code{NA} values.}
}
\value{
A logical vector.
}
\description{
Returns \code{TRUE} for \code{TRUE} elements, \code{FALSE} for \code{FALSE} elements,
and whatever is specified in \code{na} for \code{NA} items.
}
\examples{
isTrue(c(TRUE, FALSE, NA));
isTrue(c(TRUE, FALSE, NA), na=TRUE);

}
