% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biDimColors.R, R/uniDimColors.R
\name{biDimColors}
\alias{biDimColors}
\alias{uniDimColors}
\title{Create colours for a response scale for an item}
\usage{
biDimColors(start, mid, end, length, show = TRUE)

uniDimColors(start, end, length, show = TRUE)
}
\arguments{
\item{start}{Color to start with}

\item{mid}{Color in the middle, for bidimensional scales}

\item{end}{Color to end with}

\item{length}{The number of response options}

\item{show}{Whether to show the colours}
}
\value{
The colours as hex codes.
}
\description{
Create colours for a response scale for an item
}
\examples{
uniDimColors("#000000", "#00BB00", length=5, show=FALSE);
}
