% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kblXtra.R
\name{kblXtra}
\alias{kblXtra}
\title{Wrapper for kableExtra for consistent \code{ufs} table styling}
\usage{
kblXtra(
  x,
  digits = 2,
  format = "html",
  escape = FALSE,
  print = TRUE,
  viewer = FALSE,
  table.attr = "style='border:0px solid black !important;'",
  ...
)
}
\arguments{
\item{x}{The dataframe to print}

\item{digits, format, escape, table.attr}{Defaults that are
passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{print}{Wther to print the table}

\item{viewer}{Whether to show the table in the viewer}

\item{...}{Additional arguments are passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
The table, invisibly.
}
\description{
Wrapper for kableExtra for consistent \code{ufs} table styling
}
\examples{
kblXtra(mtcars);
}
