% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScales.R
\name{makeScales}
\alias{makeScales}
\title{Title}
\usage{
makeScales(data, scales, append = TRUE)
}
\arguments{
\item{data}{The dataframe containing the variables (the items).}

\item{scales}{A list of character vectors with the items in each scale,
where each vectors' name is the name of the scale.}

\item{append}{Whether to return the dataframe including the new variables
(`TRUE`), or a dataframe with only those new variables (`FALSE`).}
}
\value{
Either a dataframe with the newly created variables, or the
  supplied dataframe with the newly created variables appended.
}
\description{
Title
}
\examples{
### First generate a list with the scales
scales <- list(scale1 = c('mpg', 'cyl'), scale2 = c('disp', 'hp'));

### Create the scales and add them to the dataframe
makeScales(mtcars, scales);
}
