% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansdiamondplotjmv.h.R
\name{meansDiamondPlotjmv}
\alias{meansDiamondPlotjmv}
\title{Diamond plot: means}
\usage{
meansDiamondPlotjmv(data, items, conf.level = 95, showData = TRUE)
}
\arguments{
\item{data}{.}

\item{items}{.}

\item{conf.level}{.}

\item{showData}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a html \cr
\code{results$diamondPlot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Diamond plot: means
}
