% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiresponsejmv.h.R
\name{multiResponsejmv}
\alias{multiResponsejmv}
\title{Multi Response}
\usage{
multiResponsejmv(data, items, endorsedOption = 1)
}
\arguments{
\item{data}{.}

\item{items}{.}

\item{endorsedOption}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$table} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$table$asDF}

\code{as.data.frame(results$table)}
}
\description{
Multi Response
}
