% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFS.R
\name{nearest_vectors}
\alias{nearest_vectors}
\title{Nearest Vectors}
\usage{
nearest_vectors(df, idx = NULL, vect = NULL, n = 20, show = TRUE, ...)
}
\arguments{
\item{df}{The dataframe to plot.}

\item{vect}{Specific vectors to search from outside the dataframe.}

\item{n}{Number of neighbors to return. Defaults to 20.}

\item{show}{Shows the plot. Defaults to TRUE.}

\item{ix}{Specific index vector to search from the dataframe.}
}
\value{
Returns the indices of the n closest vectors.
}
\description{
Computes the nearest vectors to a starting vector and plots the display.
}
\examples{
nearest_vectors(senate, idx=1, n=5, main="Nearest Vectors to Index 1")
nearest_vectors(t(senate), vect=t(senate[1,]), n=5, main="Nearest Vectors to Index 1")
}
\keyword{nearest,}
\keyword{vectors}
