% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noZero.R
\name{noZero}
\alias{noZero}
\title{Remove one or more zeroes before the decimal point}
\usage{
noZero(str)
}
\arguments{
\item{str}{The character string to process.}
}
\value{
The processed string.
}
\description{
Remove one or more zeroes before the decimal point
}
\examples{
noZero("0.3");
}
\seealso{
\code{\link[=formatCI]{formatCI()}}, \code{\link[=formatR]{formatR()}}, \code{\link[=formatPvalue]{formatPvalue()}}
}
