% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelSubscales.R
\name{parallelSubscales}
\alias{parallelSubscales}
\alias{print.parallelSubscales}
\title{Split a dataset into two parallel halves}
\usage{
parallelSubscales(dat, convertToNumeric = TRUE)

\method{print}{parallelSubscales}(x, digits = 2, ...)
}
\arguments{
\item{dat}{The dataframe}

\item{convertToNumeric}{Whether to first convert all columns to numeric}

\item{x}{The object to print}

\item{digits}{The number of digits to round to}

\item{...}{Ignored.}
}
\value{
A \code{parallelSubscales} object that contains the new data frames,
and when printed shows the descriptives; or, for the print function, \code{x},
invisibly.
}
\description{
Split a dataset into two parallel halves
}
