% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFS.R
\name{plot_reduced}
\alias{plot_reduced}
\title{Plot Reduced}
\usage{
plot_reduced(df, mag = NULL, idxs = NULL, vects = NULL,
  reduce = model_pca, normalize = TRUE, gradient = "YlOrRd",
  color = "forestgreen", ...)
}
\arguments{
\item{mag}{The color scale of each of the vectors.}

\item{idxs}{Specific vectors to color from the dataframe.}

\item{vects}{Specific vectors to color from outside the dataframe.}

\item{normalize}{Automatically normalizes the magnitudes to a range from 0-1. Defaults to TRUE.}

\item{gradient}{Color gradient to use. Using a "~" reverses the direction. Defaults to "YlOrRd".}

\item{color}{Color to use for plotting points. Defaults to "forestgreen".}
}
\description{
Dimensionally reduces the columns of a dataset, plots the results, and colors by an input magnitude.
}
\examples{
plot_reduced(senate, mag=party, gradient="~RdBu", main="Senators: PCA Plot")
plot_reduced(t(senate), mag=apply(t(senate) \%*\% diag(as.vector(t(party))), 1, sum), gradient="~RdBu", main="Bills By Party")
plot_reduced(t(senate), mag=apply(senate, 2, sum), gradient="RdYlGn", main="Bills By Popularity")
plot_reduced(t(senate), main="Senate Bills: PCA Plot")
}
\keyword{plot,}
\keyword{reduced}
