% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietRemotesInstall.R
\name{quietRemotesInstall}
\alias{quietRemotesInstall}
\alias{quietGitLabUpdate}
\title{Quietly update a package from a remote repository}
\usage{
quietRemotesInstall(
  x,
  func,
  unloadNamespace = TRUE,
  dependencies = FALSE,
  upgrade = FALSE,
  quiet = TRUE,
  errorInvisible = TRUE,
  ...
)

quietGitLabUpdate(
  x,
  unloadNamespace = TRUE,
  dependencies = FALSE,
  upgrade = FALSE,
  quiet = TRUE,
  errorInvisible = TRUE,
  ...
)
}
\arguments{
\item{x}{The repository name (e.g. "\code{r-packages/ufs}")}

\item{func}{The \code{remotes} function to use}

\item{unloadNamespace}{Whether to first unload the relevant namespace}

\item{dependencies, upgrade}{Whether to install dependencies or upgrade}

\item{quiet}{Whether to suppress messages and warnings}

\item{errorInvisible}{Whether to suppress errors}

\item{...}{Additional arguments are passed on to the \code{remotes} function}
}
\value{
The result of the call to the \code{remotes} function
}
\description{
Simple wrapper for \code{remotes} functions that fail gracefully (well, don't
fail at all, just don't do what they're supposed to do) when there's no
internet connection).
}
