% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qVec.R
\name{qVec}
\alias{qVec}
\alias{qVecSum}
\title{Convenience function to quickly copy-paste a vector}
\usage{
qVec(x, fn = NULL)

qVecSum(x)
}
\arguments{
\item{x}{A string with numbers, separated by arbitrary whitespace.}

\item{fn}{An optional function to apply to the vecor before returning it.}
}
\value{
The numeric vector or result of calling the function
}
\description{
Convenience function to quickly copy-paste a vector
}
\examples{
qVec('23 	9 	11 	14 	12 	20');
}
