% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\alias{report.default}
\title{Output report from results}
\usage{
report(x, headingLevel = 3, quiet = TRUE, ...)

\method{report}{default}(x, headingLevel = 3, quiet = TRUE, ...)
}
\arguments{
\item{x}{The object to show.}

\item{headingLevel}{The level of the Markdown heading to provide; basically
the number of hashes ('\verb{#}') to prepend to the headings.}

\item{quiet}{Passed on to \code{\link[knitr:knit]{knitr::knit()}} whether it should b
chatty (\code{FALSE}) or quiet (\code{TRUE}).}

\item{...}{Passed to the specific method; for the default method, this is
passed to the print method.}
}
\description{
This method can be used to format results in a way that can directly be
included in a report or manuscript.
}
