% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeRequire.R
\name{safeRequire}
\alias{safeRequire}
\title{Load a package, install if not available}
\usage{
safeRequire(packageName, mirrorIndex = NULL)
}
\arguments{
\item{packageName}{The package}

\item{mirrorIndex}{The index of the mirror (1 is used if not specified)}
}
\description{
Load a package, install if not available
}
