% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedSubString.R
\name{sharedSubString}
\alias{sharedSubString}
\title{sharedSubString}
\usage{
sharedSubString(x, y = NULL)
}
\arguments{
\item{x}{The character vector to process.}

\item{y}{Optionally, two single values can be specified. This is probably
not useful to end users, but it's used by the function when it calls itself.}
}
\value{
A vector of length one with either the longest substring that occurs
in all values of the character vector, or NA if no overlap an be found.
}
\description{
A function to find the longest shared substring in a character vector.
}
\examples{

  sharedSubString(c("t0_responseTime", "t1_responseTime", "t2_responseTime"));
  ### Returns "_responseTime"

}
\author{
Gjalt-Jorn Peters

Maintainer: Gjalt-Jorn Peters <gjalt-jorn@userfriendlyscience.com>
}
\keyword{character}
