% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strToFilename.R
\name{strToFilename}
\alias{strToFilename}
\title{Convert a string to a safe filename}
\usage{
strToFilename(str, ext = NULL)
}
\arguments{
\item{str}{The string to convert.}

\item{ext}{Optionally, an extension to append.}
}
\value{
The string, processed to remove potentially problematic characters.
}
\description{
Convert a string to a safe filename
}
\examples{
strToFilename("this contains: illegal characters, spaces, et cetera.");
}
