% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridisPalette.R
\name{viridisPalette}
\alias{viridisPalette}
\title{Convenience function to get 2-7 color viridis palettes}
\usage{
viridisPalette(x)
}
\arguments{
\item{x}{The number of colors you want (seven at most).}
}
\value{
A vector of colours.
}
\description{
This function only exists to avoid importing the viridis package.
}
