% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_construct_export_call.R
\name{zotero_construct_export_call}
\alias{zotero_construct_export_call}
\title{Construct the URL for a Zotero export call}
\usage{
zotero_construct_export_call(
  group,
  sort = "dateAdded",
  direction = "asc",
  format = "bibtex",
  start = 0,
  limit = 100
)
}
\arguments{
\item{group}{The group ID}

\item{sort}{On which field to sort}

\item{direction}{The direction to sort in}

\item{format}{The format to export}

\item{start}{The index of the first record to return}

\item{limit}{The number of records to return}
}
\value{
The URL in a character vector.
}
\description{
This function is just a convenience function to create a simple
URL to download references from a public Zotero group. See
https://www.zotero.org/support/dev/web_api/v3/start
for details.
}
\examples{
zotero_construct_export_call(2425237);
}
