% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_get_all_items.R
\name{zotero_get_all_items}
\alias{zotero_get_all_items}
\title{Get all items in a public Zotero group}
\usage{
zotero_get_all_items(group, format = "bibtex")
}
\arguments{
\item{group}{The group ID}

\item{format}{The format to export}
}
\value{
A character vector
}
\description{
Get all items in a public Zotero group
}
\examples{
zotero_get_all_items(2425237);
}
