% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_nr_of_items.R
\name{zotero_nr_of_items}
\alias{zotero_nr_of_items}
\title{Get number of items in a public Zotero group}
\usage{
zotero_nr_of_items(group)
}
\arguments{
\item{group}{The group ID}
}
\value{
The umber of items as a numeric vector.
}
\description{
Get number of items in a public Zotero group
}
\examples{
zotero_nr_of_items(2425237);
}
