% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attenuate.d.R
\name{attenuate.d}
\alias{attenuate.d}
\title{Attenuate a Cohen's d estimate for unreliability in the continuous variable}
\usage{
attenuate.d(d, reliability)
}
\arguments{
\item{d}{The (disattenuated) value of Cohen's d}

\item{reliability}{The reliability of the measurements of the continuous variable}
}
\value{
The attenuated value of Cohen's d
}
\description{
Attenuate a Cohen's d estimate for unreliability in the continuous variable
}
\examples{
attenuate.d(.5, .8);
}
