% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findShortestInterval.R
\name{findShortestInterval}
\alias{findShortestInterval}
\title{Find the shortest interval}
\usage{
findShortestInterval(x)
}
\arguments{
\item{x}{The numeric vector.}
}
\value{
The length of the shortest interval.
}
\description{
This function takes a numeric vector, sorts it, and then finds the shortest
interval and returns its length.
}
\examples{

findShortestInterval(c(1, 2, 4, 7, 20, 10, 15));

}
\author{
Gjalt-Jorn Peters

Maintainer: Gjalt-Jorn Peters \href{mailto:gjalt-jorn@userfriendlyscience.com}{gjalt-jorn@userfriendlyscience.com}
}
\keyword{utilities}
