% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPie.R
\name{ggPie}
\alias{ggPie}
\title{A ggplot pie chart}
\usage{
ggPie(vector, scale_fill = ggplot2::scale_fill_viridis_d())
}
\arguments{
\item{vector}{The vector (best to pass a factor).}

\item{scale_fill}{The ggplot scale fill function to use for the colors.}
}
\value{
A ggplot pie chart.
}
\description{
This function creates a pie chart. Note that these are generally quite
strongly advised against, as people are not good at interpreting relative
frequencies on the basis of pie charts.
}
\note{
This function is very strongly based on the Mathematical Coffee post
at
http://mathematicalcoffee.blogspot.com/2014/06/ggpie-pie-graphs-in-ggplot2.html.
}
\examples{
ggPie(mtcars$cyl);

}
