% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTrue.R
\name{isTrue}
\alias{isTrue}
\title{More flexible version of isTRUE}
\usage{
isTrue(x, na = FALSE)
}
\arguments{
\item{x}{The vector to check for `TRUE`, `FALSE`, and `NA` values.}

\item{na}{What to return for `NA` values.}
}
\value{
A logical vector.
}
\description{
Returns `TRUE` for `TRUE` elements, `FALSE` for `FALSE` elements,
and whatever is specified in `na` for `NA` items.
}
\examples{
isTrue(c(TRUE, FALSE, NA));
isTrue(c(TRUE, FALSE, NA), na=TRUE);

}
