% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kblXtra.R
\name{kblXtra}
\alias{kblXtra}
\title{Wrapper for kableExtra for consistent \code{ufs} table styling}
\usage{
kblXtra(
  x,
  digits = 2,
  format = "html",
  escape = FALSE,
  print = TRUE,
  viewer = FALSE,
  kable_classic = FALSE,
  lightable_options = "striped",
  html_font = "\\"Arial Narrow\\", \\"Source Sans Pro\\", sans-serif",
  full_width = TRUE,
  table.attr = "style='border:0px solid black !important;'",
  ...
)
}
\arguments{
\item{x}{The dataframe to print}

\item{digits, format, escape, table.attr, lightable_options, html_font, full_width}{Defaults
that are passed to \code{\link[knitr:kable]{knitr::kable()}}}

\item{print}{Wther to print the table}

\item{viewer}{Whether to show the table in the viewer}

\item{kable_classic}{Whether to call \code{kable_classic}; otherwise,
\code{kable_styling} is called.}

\item{...}{Additional arguments are passed to \code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
The table, invisibly.
}
\description{
Wrapper for kableExtra for consistent \code{ufs} table styling
}
\examples{
kblXtra(mtcars);
}
