% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanConfInt.R
\name{meanConfInt}
\alias{meanConfInt}
\alias{print.meanConfInt}
\title{A confidence interval for the mean}
\usage{
meanConfInt(vector = NULL, mean = NULL, sd = NULL, n = NULL,
  se = NULL, conf.level = 0.95)

\method{print}{meanConfInt}(x, digits = 2, ...)
}
\arguments{
\item{vector}{A vector with raw data points - either specify this
or a mean and then either an sd and n or an se.}

\item{mean}{A mean.}

\item{sd, n}{A standard deviation and sample size; can be specified
to compute the standard error.}

\item{se}{The standard error (cna be specified instead of `sd` and `n`).}

\item{conf.level}{The confidence level of the interval.}

\item{x, digits, \dots}{Respectively the object to print, the number of digits to
round to, and any additonal arguments to pass on to the `print` function.}
}
\value{
And object with elements `input`, `intermediate`, and `output`,
where `output` holds the result in list `ci`.
}
\description{
A confidence interval for the mean
}
\examples{
meanConfInt(mean=5, sd=2, n=20);
}
